+++ src/api/trace.ts
@@ -1,23 +1,7 @@
 import { globalState } from "../core/globalstate"
-import { isComputedValue } from "../core/computedvalue"
-import { isReaction } from "../core/reaction"
 import { getAtom } from "../types/type-utils"
-import { fail, deprecated } from "../utils/utils"
-import { getMessage } from "../utils/messages"
+import { fail } from "../utils/utils"
 import { TraceMode } from "../core/derivation"
-
-function log(msg: string): string {
-    console.log(msg)
-    return msg
-}
-
-export function whyRun(thing?: any, prop?: string) {
-    deprecated("`whyRun` is deprecated in favor of `trace`")
-    thing = getAtomFromArgs(arguments)
-    if (!thing) return log(getMessage("m024"))
-    if (isComputedValue(thing) || isReaction(thing)) return log(thing.whyRun())
-    return fail(getMessage("m025"))
-}
 
 export function trace(thing?: any, prop?: string, enterBreakPoint?: boolean): void
 export function trace(thing?: any, enterBreakPoint?: boolean): void