+++ src/mobx.ts
@@ -86,7 +86,7 @@ export { action, isAction, runInAction, IActionFactory } from "./api/action"
 export { expr } from "./api/expr"
 export { toJS } from "./api/tojs"
 export { ITransformer, createTransformer } from "./api/createtransformer"
-export { whyRun, trace } from "./api/whyrun"
+export { trace } from "./api/trace"
 
 export { Iterator } from "./utils/iterable"
 export { IObserverTree, IDependencyTree, getDependencyTree, getObserverTree } from "./api/extras"